Dynamic query of selected category's words
The previous approach of simply passing down the category and using the words it contains did not work when inserting or deleting a word (it wouldn't update), so a live query is needed on every category's screen.
Having the list of words in every Category is somewhat useless and a waste now (only used for counting the number of words in the categories list), can consider to remove them.